Updating pollination to use the new kernels library.
Updating this did turn up a bug in pollination where, under certain
alphas (flight radii), the kernel could end up having even length and
width (in pixels).  This means that the kernel would not be properly
centered on the center point and was resulting in (mildly) incorrect
results.  This has now been fixed and is noted in the HISTORY.